+++ CHANGELOG.md
@@ -3,7 +3,11 @@ All notable changes to this project are documented in this file.
  
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
- 
+
+## UNRELEASED
+
+### Added
+
 ## 1.5.0 - 2017-07-17
 
 ### Added